Move API information into API file
