+++ crdtree/src/API.ts
@@ -1,9 +1,10 @@
-import {CRDTreeTransport} from "./CRDTree";
 import {FrontendPrimitive} from "./Primitive";
 import {HEAD, ROOT, ROOT_PARENT} from "./Constants";
+import {BackendChange} from "./Change";
 
 export type ID = `${string}@${number}` | HEAD | ROOT | ROOT_PARENT;
 export type Index = number | string;
+export type CRDTreeTransport<T> = BackendChange[]; // used for sending updates across the network
 
 export interface ICRDTree<T = any> {
 	render(): T;
